@@ -3,6 +3,7 @@
   "version": "1.0.0",
   "description": "Clean up regular expressions from the command line",
   "main": "index.js",
+  "type": "module",
   "bin": {
     "crgx": "./index.js"
   },
@@ -35,18 +36,17 @@
   },
   "homepage": "https://github.com/adalinesimonian/clean-regexp-cli#readme",
   "engines": {
-    "node": ">=8.0.0"
+    "node": ">=12.20"
   },
   "engineStrict": true,
   "dependencies": {
     "clean-regexp": "^1.0.0",
-    "ukaz": "^0.2.3"
+    "ukaz": "^0.3.0"
   },
   "devDependencies": {
-    "eslint-config-standard": "^10.2.1",
-    "eslint-plugin-node": "^5.1.1",
-    "eslint-plugin-standard": "^3.0.1",
-    "xo": "^0.18.2"
+    "eslint-config-standard": "^16.0.2",
+    "eslint-plugin-node": "^11.1.0",
+    "xo": "^0.40.1"
   },
   "xo": {
     "extends": [
